Rename tests to emphasize solver level tests, and now use test_case.py definitions
